requirements.txt
accelerate
bitsandbytes
datasets
huggingface-hub
langchain
langchain-cli
langchain_community
langchain-ollama
langchain-openai
llama-cpp-python
mistral_inference
mistralai
python-dotenv
sentencepiece
transformers
torch
nltk
tabulate
rouge_score
bert_score
gitpython
